Revert dev-server-init.ts and fix benchmark workflow
- Revert dev-server-init.ts extraction since the lazy loading pattern
  wasn't effective (code gets bundled anyway with static require)
- Inline the dev initialization code back into router-server.ts
- Fix benchmark workflow to use turbo build which builds dependencies
