+++ .github/workflows/test_dev_boot_time.yml
@@ -99,8 +99,7 @@ jobs:
         run: pnpm install --frozen-lockfile
 
       - name: Build Next.js (canary)
-        run: pnpm build
-        working-directory: packages/next
+        run: turbo run build --filter=next...
 
       - name: Link Next.js to test app (canary)
         run: |
@@ -149,8 +148,7 @@ jobs:
         run: pnpm install --frozen-lockfile
 
       - name: Build Next.js (PR)
-        run: pnpm build
-        working-directory: packages/next
+        run: turbo run build --filter=next...
 
       - name: Link Next.js to test app (PR)
         run: |
@@ -354,8 +352,7 @@ jobs:
         run: pnpm install
 
       - name: Build Next.js
-        run: pnpm build
-        working-directory: packages/next
+        run: turbo run build --filter=next...
 
       - name: Run bytecode cache unit tests
         run: pnpm jest test/unit/bytecode-cache.test.ts --runInBand
